Add files via upload
Given a string, display only those characters which are present at an even index number. Given a string and an integer number n, remove characters from a string starting from zero up to n and return a new string. For example, removeChars("pynative", 4) so output must be tive. Note: n must be less than the length of the string.